requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour-science==0.4.4
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
Django==5.0.1
django-cors-headers==4.3.1
django-rest-knox==4.2.0
django-tenants==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dynaconf==3.2.4
fonttools==4.47.2
hvac==2.1.0
idna==3.6
imageio==2.33.1
kiwisolver==1.4.5
knox==0.1.14
loguru==0.7.2
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
scipy==1.12.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
types-pkg-resources==0.1.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
validators==0.22.0
win32-setctime==1.1.0
